+++ src/Mzinga/Engine/Engine.cs
@@ -609,6 +609,8 @@ namespace Mzinga.Engine
         private void OptionsList()
         {
             OptionsGet("MaxBranchingFactor");
+            OptionsGet("QuiescentSearchMaxDepth");
+            OptionsGet("PrincipalVariationMaxDepth");
             OptionsGet("MaxHelperThreads");
             OptionsGet("PonderDuringIdle");
             OptionsGet("TranspositionTableSizeMB");
@@ -634,6 +636,14 @@ namespace Mzinga.Engine
                 case "MaxBranchingFactor":
                     DefaultConfig.GetMaxBranchingFactorValue(out _, out defaultValue, out _);
                     Config.GetMaxBranchingFactorValue(out type, out value, out values);
+                    break;
+                case "QuiescentSearchMaxDepth":
+                    DefaultConfig.GetQuiescentSearchMaxDepthValue(out _, out defaultValue, out _);
+                    Config.GetQuiescentSearchMaxDepthValue(out type, out value, out values);
+                    break;
+                case "PrincipalVariationMaxDepth":
+                    DefaultConfig.GetPrincipalVariationMaxDepthValue(out _, out defaultValue, out _);
+                    Config.GetPrincipalVariationMaxDepthValue(out type, out value, out values);
                     break;
                 case "MaxHelperThreads":
                     DefaultConfig.GetMaxHelperThreadsValue(out _, out defaultValue, out _);
@@ -683,6 +693,14 @@ namespace Mzinga.Engine
                     Config.ParseMaxBranchingFactorValue(value);
                     refreshAI = true;
                     resetCaches = true;
+                    break;
+                case "QuiescentSearchMaxDepth":
+                    Config.ParseQuiescentSearchMaxDepthValue(value);
+                    refreshAI = true;
+                    resetCaches = true;
+                    break;
+                case "PrincipalVariationMaxDepth":
+                    Config.ParsePrincipalVariationMaxDepthValue(value);
                     break;
                 case "MaxHelperThreads":
                     Config.ParseMaxHelperThreadsValue(value);